Added Windows version
Added Windows version. It seems that I should probably redesign the way
the front end works - it should class-based, for one. Furthermore, it
looks like the method I am using in the Windows version to activate the
translation should be improved, as it is impossible to hide the text
being inputted. Instead, the use of buttons and/or a text input would be
preferable,